% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{inline}
\alias{inline}
\title{Container for inline widgets}
\usage{
inline(..., class = "mb-1")
}
\arguments{
\item{...}{Unnamed arguments: Inline widgets (such as \code{\link[=inline_text]{inline_text()}}),
character strings, or \link[htmltools:tag]{HTML tags} that will appear next
to each other in a line or paragraph. These are pasted together with no
spaces between them, so add extra spaces to your character strings if
needed. Named arguments (e.g. \code{style}) are additional attributes for the
HTML \link[htmltools:div]{div} tag wrapping the line.}

\item{class}{Extra classes to apply to the line. The default, \code{"mb-1"}, is a
Bootstrap 5 class that adds a small amount of margin to the bottom of
the line. You can use \code{"mb-0"} through \code{"mb-5"}, other
\href{https://getbootstrap.com/docs/5.3/utilities/spacing/}{Bootstrap 5 spacing classes},
or anything else. For multiple classes, provide one space-separated
string.}
}
\value{
An HTML element to be included in your Shiny UI definition.
}
\description{
Wrapper for a line (or paragraph) containing a mix of explanatory text and
inshiny inline widgets.
}
\section{Note on rendering}{


Inline widgets from inshiny are typically displayed with borders around them
to show that they are editable. For the borders to display correctly, inshiny
includes some custom CSS to make sure the borders work inside common
\link[bslib:bslib-package]{bslib} containers such as \link[bslib:card]{bslib::card} and
\link[bslib:accordion_panel]{bslib::accordion_panel}. However, if you place inline widgets inside a
custom container that has its own opaque background (e.g. a \code{div} with
\code{background-color} set), the border around widgets may disappear. To fix
this, add the CSS class \code{"inshiny-bg"} to the container:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tags$div(class = "inshiny-bg", style = "background-color: white;",
    inline("Enter a value: ", inline_text("val", "hello"))
)
}\if{html}{\out{</div>}}
}

\examples{
ui <- bslib::page_fixed(
    shiny::h1("Hello!"),
    inline("My name is ", inline_text("myname", "Sisyphus"), "."),
    inline("Please enter your name carefully.", style = "font-weight:bold")
)
}
