% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inshiny-package.R
\docType{package}
\name{inshiny-package}
\alias{inshiny}
\alias{inshiny-package}
\title{inshiny: Compact Inline Widgets for 'shiny' Apps}
\description{
Provides a basic set of compact widgets for 'shiny' apps which occupy less space and can appear inline with surrounding text.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicholasdavies/inshiny}
  \item \url{https://nicholasdavies.github.io/inshiny/}
  \item Report bugs at \url{https://github.com/nicholasdavies/inshiny/issues}
}

}
\author{
\strong{Maintainer}: Nick Davies \email{nicholas.davies@lshtm.ac.uk}

Authors:
\itemize{
  \item Nick Davies \email{nicholas.davies@lshtm.ac.uk}
}

}
\keyword{internal}
